the order 'extra field' tables that were previously located directly inside models/ordermodel.py are now extensible.
