+++ shop/models/defaults/bases.py
@@ -426,3 +426,82 @@ class BaseOrderItem(models.Model):
         if self.product:
             self.product_name = self.product.get_name()
         super(BaseOrderItem, self).save(*args, **kwargs)
+
+
+#==============================================================================
+# Order Extras
+#==============================================================================
+
+
+class BaseOrderExtraInfo(models.Model):
+    """
+    A holder for extra textual information to attach to this order.
+    """
+    order = models.ForeignKey(get_model_string('Order'), related_name="extra_info",
+            verbose_name=_('Order'))
+    text = models.TextField(verbose_name=_('Extra info'))
+
+    class Meta(object):
+        abstract = True
+        app_label = 'shop'
+        verbose_name = _('Order extra info')
+        verbose_name_plural = _('Order extra info')
+
+
+class BaseExtraOrderPriceField(models.Model):
+    """
+    This will make Cart-provided extra price fields persistent since we want
+    to "snapshot" their statuses at the time when the order was made
+    """
+    order = models.ForeignKey(get_model_string('Order'), verbose_name=_('Order'))
+    label = models.CharField(max_length=255, verbose_name=_('Label'))
+    value = CurrencyField(verbose_name=_('Amount'))
+
+    # Does this represent shipping costs?
+    is_shipping = models.BooleanField(default=False, editable=False,
+            verbose_name=_('Is shipping'))
+
+    class Meta(object):
+        abstract = True
+        app_label = 'shop'
+        verbose_name = _('Extra order price field')
+        verbose_name_plural = _('Extra order price fields')
+
+
+class BaseExtraOrderItemPriceField(models.Model):
+    """
+    This will make Cart-provided extra price fields persistent since we want
+    to "snapshot" their statuses at the time when the order was made
+    """
+    order_item = models.ForeignKey(get_model_string('OrderItem'), verbose_name=_('Order item'))
+    label = models.CharField(max_length=255, verbose_name=_('Label'))
+    value = CurrencyField(verbose_name=_('Amount'))
+
+    class Meta(object):
+        abstract = True
+        app_label = 'shop'
+        verbose_name = _('Extra order item price field')
+        verbose_name_plural = _('Extra order item price fields')
+
+
+class BaseOrderPayment(models.Model):
+    """
+    A class to hold basic payment information. Backends should define their own
+    more complex payment types should they need to store more informtion
+    """
+    order = models.ForeignKey(get_model_string('Order'), verbose_name=_('Order'))
+    amount = CurrencyField(verbose_name=_('Amount'))# How much was payed with this particular transfer
+    transaction_id = models.CharField(max_length=255,
+            verbose_name=_('Transaction ID'),
+            help_text=_("The transaction processor's reference"))
+    payment_method = models.CharField(max_length=255,
+            verbose_name=_('Payment method'),
+            help_text=_("The payment backend use to process the purchase"))
+
+    class Meta(object):
+        abstract = True
+        app_label = 'shop'
+        verbose_name = _('Order payment')
+        verbose_name_plural = _('Order payments')
+
+
